Prevented use of core logs outside engine DLL
- Prroject generation scripts check for vulkan_sdk env variable

+++ GenerateProjectsMac.sh
@@ -1,5 +1,12 @@
 #!/bin/bash
 
+if [ -z "$VULKAN_SDK" ]; then
+    echo "Error: VULKAN_SDK environment variable is not set."
+    echo "Please ensure Vulkan SDK is installed and the environment variable is configured."
+    read -p "Press any key to exit..." -n1
+    exit 1
+fi
+
 rm -rf bin
 rm -rf bin-int
